pyproject.toml
[tool.poetry]
name = "donaci-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.1"
djangorestframework = "^3.9"
gunicorn = "^19.9"
django_heroku = "^0.3.1"
whitenoise = "^4.1"
Pillow = "^5.3"
psycopg2-binary = "^2.7"

[tool.poetry.dev-dependencies]
pyflakes = "^2.0"
jedi = "^0.13.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
